==> Code/step3_sequence_chisquare.ipynb <==
# NSDUH Drug Sequence Analysis Part 3:  Chi-Square Analysis
## Matthew J. Beattie
## University of Oklahoma
__February 5, 2021__

This notebook does a deeper dive on Chi-Square analysis of the clusters with regard to their demographic categories. 

"""
Import python modules
"""
import pandas as pd
import numpy as np
import copy
import os
import pathlib, itertools
import time as timelib
import matplotlib.pyplot as plt
from sklearn.pipeline import Pipeline
from sklearn.base import BaseEstimator, TransformerMixin
import random
from sklearn_extra import cluster as cs
from sklearn.cluster import KMeans
import pickle
import json
import pathutils as pu
import scipy.stats as stats
from math import dist
import mlflow

# Parallelize pandas with modin and dash
#import modin.pandas as pd

HOME_DIR = pathlib.Path.home()
CW_DIR = pathlib.Path.cwd()

FIGW = 12
FIGH = 5
FONTSIZE = 8
FIGURESIZE = (FIGW,FIGH)

plt.rcParams['figure.figsize'] = (FIGW, FIGH)
plt.rcParams['font.size'] = FONTSIZE

plt.rcParams['xtick.labelsize'] = FONTSIZE
plt.rcParams['ytick.labelsize'] = FONTSIZE

random.seed(660806)


### Import Data
Import data from pickle files and prepare for analysis

# Set parameters
datapath = 'C:/Users/mjbea/OneDrive/GitHub/abuse_sequence/Data3/'
workingpath = 'C:/Users/mjbea/OneDrive/GitHub/abuse_sequence/Code3/'
outpath = 'C:/Users/mjbea/OneDrive/GitHub/abuse_sequence/Output3/'
year = '2016_2017_2018_2019'
jsondict = datapath + 'NSDUH_field.json'
n_clusters = 11

# Setup filenames
clustpkl = workingpath + 'Kmeans_' + str(n_clusters) + 'clust_' + str(year) + '_nonullpath_clust.pkl'
demogpkl = workingpath + 'Kmeans_' + str(n_clusters) + 'clust_' + str(year) + '_nonullpath_demog.pkl'
modpkl = workingpath + 'Kmeans_' + str(n_clusters) + 'clust_' + str(year) + '_nonullpath_model.pkl'

# Read in cluster and demographic information and the model
dfclust = pd.read_pickle(clustpkl)
dfdemog = pd.read_pickle(demogpkl)
model = pickle.load(open(modpkl, 'rb'))

# Drug name and indices are called by a user-defined function.
ident, rawafuvals, afuvals, drugnames, drugorder, drugnums, drugposition, startdemog, demographics = pu.surveyvars(year)

# Decode dictionaries for NSDUH variables:
f1 = open(jsondict, 'r')
nsduhDecoder = json.load(f1)
f1.close()

strings = 0
ints = 0
selse = 0
for i in range(0,len(dfclust['RESPID'])):
    if type(dfclust['RESPID'][i])==str:
        strings += 1
    elif type(dfclust['RESPID'][i])==int:
        ints += 1
    else:
        selse += 1
print(strings, ints, selse)

model.inertia_

### Review data and key variables

# Show cluster dataframe
dfclust['HEROIN_USER'] = dfclust.apply(lambda row: 'Heroin User' if (row['AFUVECT'][8] > 0 and row['AFUVECT'][8] < 991) else 'Not Heroin User', axis=1)
dfclust.head()

dfclust['HEROIN_USER'].value_counts()

# Show demographic dataframe
dfdemog.head()

# Show category values for demographic variables
nsduhDecoder

catcrossnorm = pd.crosstab(index=dfclust['labels'], columns=['Percentage'], normalize='columns')
catcrossnorm

# Set output file variables
jsonout = outpath + 'Kmeans_' + str(n_clusters) + 'clust_' + str(year) + '_alldata_clust_stats.json'
outfile = outpath + 'Kmeans_' + str(n_clusters) + 'clust_' + str(year) + '_chi2_output.txt'
outfile2 = outpath + 'Kmeans_' + str(n_clusters) + 'clust_' + str(year) + '_crosstab_latex.txt'
csvout = outpath + 'Kmeans_' + str(n_clusters) + 'clust_' + str(year) + '_ranked_chi2.csv'

# Calculate medoids from cluster model and store center and medoids into a dataframe and dictionary for output to json
clustsmrydf = pd.DataFrame()
tempdict = {}
medoids = []
for j in range(0, len(model.cluster_centers_)):
    clustcenter = model.cluster_centers_[j]
    clust = dfclust.loc[dfclust['labels']==j,:]
    intcenter = np.array(clustcenter, dtype=int)
    if clust.shape[0] > 0:
        clust['distfrommean'] = clust.apply(lambda row: dist(np.array(row['AFUVECT']), clustcenter), axis=1)
        count = clust.shape[0]
        clustfrac = round(clust.shape[0]/dfclust.shape[0],5)
        medoid = clust.loc[clust['distfrommean'].idxmin()]['AFUVECT']
        medoids.append(medoid)
        path = pu.uncodevect(medoid, drugnames)
        clustsmrydf = clustsmrydf.append({'clustnum': j, 'count': count, 'clustfrac': clustfrac, 'center': clustcenter, 
                                          'intcenter': intcenter, 'medoid': medoid, 'path': path}, ignore_index=True)
        tempdict.update({j: {'count': count, 'clustfrac': clustfrac, 'center': clustcenter.tolist(), 'intcenter': intcenter.tolist(), 'medoid': medoid, 'path': path}})
    else:
        clustsmrydf = clustsmrydf.append({'clustnum': j, 'count': 0, 'clustfrac': 0, 'center': clustcenter, 'intcenter': intcenter, 'medoid': 'NA', 'path': 'NA'}, ignore_index=True)
        tempdict.update({j: {'count': 0, 'clustfrac': 0, 'center': clustcenter.tolist(), 'intcenter': intcenter.tolist(), 'medoid': 'NA', 'path': 'NA'}})

# Save cluster dictionary to json file
with open(jsonout,'w') as cluststats_dumped:
    json.dump(tempdict,cluststats_dumped, indent = 4, sort_keys = True)

cluststats_dumped.close()

# Print cluster summary output leave file open for crosstabs
f = open(outfile, 'w')

# Open latex output file
f2 = open(outfile2, 'w')

print('\n******\nCLUSTER CENTERS\n******', file=f)
print(model.cluster_centers_, file=f)
print('\n******\n INTEGER CLUSTER CENTERS\n******', file=f)
print(np.array(model.cluster_centers_, dtype='int'), file=f)
print('\n******\n CLUSTER MEDOIDS\n******\n', file=f)
print(medoids, file=f)
print('\n******\nCLUSTER ITERATIONS\n******', file=f)
print(model.n_iter_, file=f)
print('\n******\nCLUSTER INERTIA\n******', file=f)
print(model.inertia_, file=f)

# Print out cluster pathways and save translation to a list
print('\n******\nKNN CLUSTER PATHWAYS\n******', file=f)
centerpaths = []
i = 0
for i in range(0, len(model.cluster_centers_)):
    path = pu.uncodevect(model.cluster_centers_[i], drugnames)
    print('Cluster ', i, '\n', path, file=f)
    centerpaths.append(path)
    
# Print out cluster pathways and save translation to a list
print('\n******\nKNN PSUDEO-CLUSTER PATHWAYS\n******', file=f)
centerpaths = []
i = 0
for i in range(0, len(model.cluster_centers_)):
    path = pu.uncodevect(model.cluster_centers_[i], drugnames)
    path = [item for item in path if item[1] < 99]
    print('Cluster ', i, '\n', path, file=f)
    centerpaths.append(path)
    
print('\n******\nPSUEDO-MEDOID CLUSTER PATHWAYS\n******', file=f)
centerpaths = []
i = 0
for i in range(0, len(model.cluster_centers_)):
    path = pu.uncodevect(medoids[i], drugnames)
    print('Cluster ', i, '\n', path, file=f)
    centerpaths.append(path)
    
print('\n******\nCLUSTER RAW COUNTS\n******', file=f)
print(dfclust['labels'].value_counts(), file=f)

print('\n******\nCLUSTER PERCENTAGES\n******', file=f)
print(pd.crosstab(index=dfclust['labels'], columns=['Percentage'], normalize='columns'), file=f)
print('Fraction of respondents by cluster:', file=f2)
print(pd.crosstab(index=dfclust['labels'], columns=['Percentage'], normalize='columns').to_latex(), file=f2)

print('\n******\nEND OF SECTION\n******\n', file=f)


### Generate categorical crosstabs
The demographic dataset needs to be converted to crosstabs for chi-square analysis

# Convert fields in the demographic database to category variables.
catvars = copy.deepcopy(demographics)
catvars.remove('AGE2')
catvars.remove('RESPID')

# Create a new dataframe for translation of categorical values into
# reader friendly values from the NSDUH dictionary.
# Youth dictionary values are not considered.
tmpdf = copy.deepcopy(dfdemog)
pipeprep = Pipeline([('makecats', pu.MakeCat(catvars))])
tmpdf = pipeprep.transform(tmpdf)

for fieldname in tmpdf.columns:
    newvallist = []
    if fieldname != 'RESPID' and fieldname != 'labels':
        for key, value in nsduhDecoder[fieldname]['values'].items():
            if value != "Youth":
                newvallist.append(value)
        tmpdf[fieldname].cat.categories = newvallist

### Run Chi-square analysis
Run chi2 for each demographic variable.  Calculate chi2 values for the individual cells and store their values along with the fraction of the total table chi2.  Call this last statistic _INFLUENCE_.

# Initialize individual chi2 dataframe
indchi2df = pd.DataFrame(columns = ['TABLE', 'VALUE', 'CLUSTER', 'OBSERVED', 'EXPECTED', 'DIFF', 'INDCHI2', 'INFLUENCE'])

print('\n*****\nCROSSTABS AND CHISQUARE TEST RESULTS\n*****\n', file=f)
for item in catvars:
    catcross = pd.crosstab(index=tmpdf['labels'], columns=tmpdf[item])
    catcrossnorm = pd.crosstab(index=tmpdf['labels'], columns=tmpdf[item], normalize='columns')
    chi2, p, dof, ex = stats.chi2_contingency(catcross)
    print('\n******', item, '******\n', file=f)
    print('CROSSTAB COLUMN PERCENTAGES', file=f)
    print(catcrossnorm, '\n', file=f)
    print(catcrossnorm.to_latex(index=False), '\n', file=f2)
    print('CHI SQUARE RESULTS', file=f)
    print('CHI2 value: ', chi2, 'p-value: ', p, 'Degrees of freedom: ', dof, '\n', file=f)

    # Write individual chi2 to table
    rowcnt, colcnt = catcross.shape[0], catcross.shape[1]
    for i in range(0, rowcnt):
        newrow = []
        for j in range(0, colcnt):
            table = item
            value = catcross.columns[j]
            observed = catcross.loc[i][j]
            expected = ex[i,j]
            diff = observed - expected
            indchi2 = (catcross.loc[i][j] - ex[i,j])**2/ex[i,j]
            dftmp = {'TABLE': table, 'VALUE': value, 'CLUSTER': i, 'OBSERVED': observed, 'EXPECTED': expected,
                     'DIFF': diff, 'INDCHI2': indchi2, 'INFLUENCE': indchi2/chi2}
            indchi2df = indchi2df.append(dftmp, ignore_index = True)
            

# Duplicate process for heroin use
catcross = pd.crosstab(index=dfclust['labels'], columns=dfclust['HEROIN_USER'])
catcrossnorm = pd.crosstab(index=dfclust['labels'], columns=dfclust['HEROIN_USER'], normalize='columns')
chi2, p, dof, ex = stats.chi2_contingency(catcross)
print('\n******', item, '******\n', file=f)
print('HEROIN USE CROSSTAB COLUMN PERCENTAGES', file=f)
print(catcrossnorm, '\n', file=f)
print(catcrossnorm.to_latex(index=False), '\n', file=f2)
print('CHI SQUARE RESULTS', file=f)
print('CHI2 value: ', chi2, 'p-value: ', p, 'Degrees of freedom: ', dof, '\n', file=f)

# Write individual chi2 to table for heroin use
rowcnt, colcnt = catcross.shape[0], catcross.shape[1]
for i in range(0, rowcnt):
    newrow = []
    for j in range(0, colcnt):
        table = 'HEROIN_USE'
        value = catcross.columns[j]
        observed = catcross.loc[i][j]
        expected = ex[i,j]
        diff = observed - expected
        indchi2 = (catcross.loc[i][j] - ex[i,j])**2/ex[i,j]
        dftmp = {'TABLE': table, 'VALUE': value, 'CLUSTER': i, 'OBSERVED': observed, 'EXPECTED': expected,
                 'DIFF': diff, 'INDCHI2': indchi2, 'INFLUENCE': indchi2/chi2}
        indchi2df = indchi2df.append(dftmp, ignore_index = True)


indchi2df.head()

### Rank the entries in the chi-square crosstab
To find the individual cells in the chi2 crosstab with the most influence, rank them by chi2 score.

indchi2df['rank'] = indchi2df.groupby('TABLE')['INDCHI2'].rank(method="first", ascending=False)
indchi2df.to_csv(csvout)
pd.set_option('max_rows', 100)
indchi2df.head(100)

pd.reset_option('max_rows')

topindchi2 = indchi2df[indchi2df['rank']<=5]
topindchi2.head(100)

# Print out top individual chi2 to output
print(topindchi2, file=f)
print(topindchi2.to_latex(index=False), file=f2)

# Close print files
f.close()
f2.close()

# Log runs to mlflow
experiment_name = "Chi-Square Analysis"
mlflow.set_experiment(experiment_name)

with mlflow.start_run():
    mlflow.log_param("n_clusters", n_clusters)
    mlflow.log_param("Observations", 'All')
    
    mlflow.log_artifact(jsonout)
    mlflow.log_artifact(csvout)
    mlflow.log_artifact(outfile)
    mlflow.log_artifact(outfile2)
   
    # End mlflow run
    mlflow.end_run()
